Fix some of the warnings as reported by MS VC++ 2010 Express:  * bitify would need more thought
git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@491 70805675-57b6-444b-bb00-18ecffcb4be8

+++ cru/guessio/cru_1901_2006.h
@@ -212,7 +212,7 @@ private:
 
 	void bitify(unsigned char buf[4],double fval,double offset,double scalar) {
 
-		long ival=(fval-offset)/scalar+0.5;
+		long ival = (long)((fval-offset)/scalar + 0.5);
 		buf[0]=ival/0x1000000;
 		ival-=buf[0]*0x1000000;
 		buf[1]=ival/0x10000;
